feat: Add MHash struct and hash function
Add a new struct called MHash that implements the Hasher trait. This struct is used to calculate a hash value based on a given seed and a value. The hash function takes into account the previous byte and performs a series of calculations to generate the hash value.

Also, add the hasher module to the supplementary module, which contains various utility modules.